[INLONG-9600][Agent] Adjust the sins directory for code consistency
