+++ inlong-agent/agent-plugins/src/main/java/org/apache/inlong/agent/plugin/sinks/PulsarSink.java
@@ -19,15 +19,11 @@ package org.apache.inlong.agent.plugin.sinks;
 
 import org.apache.inlong.agent.common.AgentThreadFactory;
 import org.apache.inlong.agent.conf.AgentConfiguration;
-import org.apache.inlong.agent.conf.JobProfile;
+import org.apache.inlong.agent.conf.InstanceProfile;
 import org.apache.inlong.agent.message.BatchProxyMessage;
-import org.apache.inlong.agent.message.EndMessage;
-import org.apache.inlong.agent.message.PackProxyMessage;
-import org.apache.inlong.agent.message.ProxyMessage;
 import org.apache.inlong.agent.metrics.audit.AuditUtils;
 import org.apache.inlong.agent.plugin.Message;
 import org.apache.inlong.agent.utils.AgentUtils;
-import org.apache.inlong.agent.utils.ThreadUtils;
 import org.apache.inlong.common.msg.InLongMsg;
 import org.apache.inlong.common.pojo.dataproxy.MQClusterInfo;
 
@@ -115,7 +111,7 @@ public class PulsarSink extends AbstractSink {
     private boolean asyncSend;
 
     @Override
-    public void init(JobProfile jobConf) {
+    public void init(InstanceProfile jobConf) {
         super.init(jobConf);
         // agentConf
         sendQueueSize = agentConf.getInt(PULSAR_SINK_SEND_QUEUE_SIZE, DEFAULT_SEND_QUEUE_SIZE);
@@ -149,41 +145,11 @@ public class PulsarSink extends AbstractSink {
         pulsarSenders = new ArrayList<>();
         initPulsarSender();
         EXECUTOR_SERVICE.execute(sendDataThread());
-        EXECUTOR_SERVICE.execute(flushCache());
     }
 
     @Override
-    public void write(Message message) {
-        try {
-            if (message != null) {
-                if (!(message instanceof EndMessage)) {
-                    ProxyMessage proxyMessage = new ProxyMessage(message);
-                    // add proxy message to cache.
-                    cache.compute(proxyMessage.getBatchKey(),
-                            (s, packProxyMessage) -> {
-                                if (packProxyMessage == null) {
-                                    packProxyMessage =
-                                            new PackProxyMessage(jobInstanceId, jobConf, inlongGroupId, inlongStreamId);
-                                    packProxyMessage.generateExtraMap(proxyMessage.getDataKey());
-                                    packProxyMessage.addTopicAndDataTime(topic, System.currentTimeMillis());
-                                }
-                                // add message to package proxy
-                                packProxyMessage.addProxyMessage(proxyMessage);
-                                return packProxyMessage;
-                            });
-                    // increment the count of successful sinks
-                    sinkMetric.sinkSuccessCount.incrementAndGet();
-                } else {
-                    // increment the count of failed sinks
-                    sinkMetric.sinkFailCount.incrementAndGet();
-                }
-            }
-        } catch (Exception e) {
-            LOGGER.error("write message to Proxy sink error", e);
-        } catch (Throwable t) {
-            ThreadUtils.threadThrowableHandler(Thread.currentThread(), t);
-        }
-
+    public boolean write(Message message) {
+        return true;
     }
 
     @Override
@@ -203,45 +169,8 @@ public class PulsarSink extends AbstractSink {
         }
     }
 
-    private boolean sinkFinish() {
-        return cache.values().stream().allMatch(PackProxyMessage::isEmpty) && pulsarSendQueue.isEmpty();
-    }
-
-    /**
-     * flush cache by batch
-     *
-     * @return thread runner
-     */
-    private Runnable flushCache() {
-        return () -> {
-            LOGGER.info("start flush cache thread for {} ProxySink", inlongGroupId);
-            while (!shutdown) {
-                try {
-                    cache.forEach((batchKey, packProxyMessage) -> {
-                        BatchProxyMessage batchProxyMessage = packProxyMessage.fetchBatch();
-                        if (batchProxyMessage != null) {
-                            try {
-                                sendQueueSemaphore.acquire();
-                                pulsarSendQueue.put(batchProxyMessage);
-                                LOGGER.info("send group id {}, message key {},with message size {}, the job id is {}, "
-                                        + "read source is {} sendTime is {}", inlongGroupId, batchKey,
-                                        batchProxyMessage.getDataList().size(), jobInstanceId, sourceName,
-                                        batchProxyMessage.getDataTime());
-                            } catch (Exception e) {
-                                sendQueueSemaphore.release();
-                                LOGGER.error("flush data to send queue", e);
-                            }
-                        }
-                    });
-                } catch (Exception ex) {
-                    LOGGER.error("error caught", ex);
-                } catch (Throwable t) {
-                    ThreadUtils.threadThrowableHandler(Thread.currentThread(), t);
-                } finally {
-                    AgentUtils.silenceSleepInMs(batchFlushInterval);
-                }
-            }
-        };
+    public boolean sinkFinish() {
+        return true;
     }
 
     /**
